Rename test project to .Tests so it is a nested namespace.
