--- c-sharp/TicTacToeTest/TicTacToeTest.csproj
@@ -1,19 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <ProjectGuid>{322E6DF6-4431-4B1B-9E9C-A631342A2DEB}</ProjectGuid>
-    <RootNamespace>TicTacToeTest</RootNamespace>
-    <AssemblyName>TicTacToeTest</AssemblyName>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.3.0" />
-    <PackageReference Include="MSTest.TestAdapter" Version="1.1.18" />
-    <PackageReference Include="MSTest.TestFramework" Version="1.1.18" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\TicTacToe\TicTacToe.csproj" />
-  </ItemGroup>
-</Project>
